refactor forecast component
