--- src/Components/Forecast.css
@@ -1,4 +1,0 @@
-.weather-forecast {
-  text-align: center;
-  border: 1px solid red;
-}
