+++ src/Components/WeatherApp.css
@@ -199,6 +199,7 @@ body {
 }
 
 .weather-forecast h2 {
+  text-align: left;
   font-size: 14px;
   margin: 8px 16px;
   font-weight: 600;
